Introduce `mocha` test coverage
This commit is the output of `mocha init test`.

This boilerplate is required to run the `mocha` command to execute the
test suite.

Unfortunately, `mocha init`'s output does not fully account for being
executed in a Node application that uses `global` as the global context
instead of `window`:

```
~/src/why-test-web-app master*
❯ mocha
/Users/seanpdoyle/src/why-test-web-app/test/mocha.js:12413
window.Mocha = Mocha;
^

ReferenceError: window is not defined
```

In the interest in tracking scaffolded code separately from code that
we've authored, the fix will be included in a follow-up commit.