Change server test verification
Our Feature tests give us confidence that the contents of a message and
its author are rendered onto the page.

This commit modifies a feature test to verify that a `Message` record is
created as a side-effect.

Additionally, we assert that invalid parameters result in a response
with a [422 HTTP Status][docs].

[docs]: https://httpstatuses.com/422