Fix `bin/mocha-test` to handle multiple files
There is currently a subtle bug in our `mocha-test` script.

The `mocha` command cannot accept a single argument String containing
the names of the test files to run. Instead, it expects the file names
delimited by whitespace.

This commit invokes the `mocha` command included in `node_modules/`, and
no longer wraps the test file names as a single String argument.