Rename `test/requests` to `test/routes`
Having a `requests` directory in our tests is borrowed from RSpec and
Rails.

It made sense for testing an Express application, since the server
components under test are responsible for handling HTTP requests.

In an effort for learners to more easily grasp "The Upside Down", let's
call the directory `routes` to more strongly emphasize that the tests
are covering our Express routes.