Add `jsdom` for Server Test assertions
This commit adds [`jsdom@9.12.0`][jsdom] as a dependency.

The Learning Environment already makes `9.12.0` available to us.

Before this commit, server tests were using raw String comparison in its
verification phases. This yielded verbose assertion failures that were
difficult to read and understand.

The goal of introducing JSDOM is to provide server-level tests with the
same CSS selector support that we make use of in our feature tests.

[jsdom]: https://github.com/tmpvar/jsdom/tree/9.12.0#jsdom