Route test asserts redirection after creation
Instead of duplicating coverage provided at the feature level, the test
covering `POST /messages` asserts that the response will redirect to the
`index` route, mounted to `GET /`.

Support for this change was test-driven by splitting the route test's
Happy Path into two separate `it` blocks, with fewer assertions in each.

We could use this change to draw learner attention to the fact that
there are two _separate_ side-effects we're asserting here, which can
improve each test's _readability_ and our suite's _completeness_.